+++ codegen/src/main/scala/overflowdb/codegen/Helpers.scala
@@ -192,10 +192,39 @@ object Helpers {
     }
   }
 
-  /** in theory there can be multiple candidates - we're just returning one of those for now */
-  def lowestCommonAncestor(nodes: Seq[AbstractNodeType]): Option[AbstractNodeType] = {
-    LowestCommonAncestors1(nodes.toSet)(_.extendzRecursively.toSet).headOption
+  // TODO return AbstractNodeType - need DefaultNodeTypes.StoredNode: AbstractNodeType
+  def deriveCommonRootType(neighborNodeInfos: Set[AbstractNodeType]): String = {
+    lowestCommonAncestor(neighborNodeInfos)
+      .orElse(findSharedRoot(neighborNodeInfos))
+      .map(_.className)
+      .getOrElse(DefaultNodeTypes.StoredNodeClassname)
   }
+
+  /** in theory there can be multiple candidates - we're just returning one of those for now */
+  def lowestCommonAncestor(nodes: Set[AbstractNodeType]): Option[AbstractNodeType] = {
+    LowestCommonAncestors1(nodes)(_.extendzRecursively.toSet).headOption
+  }
+
+  /** from the given node types, find one that is part of the complete type hierarchy of *all* other node types */
+  def findSharedRoot(nodeTypes: Set[AbstractNodeType]): Option[AbstractNodeType] = {
+    if (nodeTypes.size == 1) {
+      Some(nodeTypes.head)
+    } else if (nodeTypes.size > 1) {
+      // trying to keep it deterministic...
+      val sorted = nodeTypes.toSeq.sortBy(_.className)
+      val (first, otherNodes) = (sorted.head, sorted.tail)
+      completeTypeHierarchy(first).find { candidate =>
+        otherNodes.forall { otherNode =>
+          completeTypeHierarchy(otherNode).contains(candidate)
+        }
+      }
+    } else {
+      None
+    }
+  }
+
+  def completeTypeHierarchy(node: AbstractNodeType): Seq[AbstractNodeType] =
+    node +: node.extendzRecursively
 
 }
