+++ codegen/src/test/scala/overflowdb/schema/SchemaTest.scala
@@ -2,7 +2,6 @@ package overflowdb.schema
 
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
-import overflowdb.codegen.LowestCommonAncestors1
 
 class SchemaTest extends AnyWordSpec with Matchers {
 
@@ -48,39 +47,15 @@ class SchemaTest extends AnyWordSpec with Matchers {
       val baseNodeA = new NodeBaseType(name = "BASE_A", comment = None, SchemaInfo.Unknown)
       val nodeA1 = new NodeType(name = "A1", comment = None, SchemaInfo.Unknown).extendz(baseNodeA)
 
-      val nodeInfos = neighborInfoWith(Seq(nodeA1, baseNodeA)).nodeInfos.map(_.neighborNode).toSet
-      val lcaMaybe = LowestCommonAncestors1(nodeInfos)(_.extendzRecursively.toSet).headOption
-      println(lcaMaybe)
+//      val nodeInfos = neighborInfoWith(Seq(nodeA1, baseNodeA)).nodeInfos.map(_.neighborNode).toSet
+//      val lcaMaybe = LowestCommonAncestors1(nodeInfos)(_.extendzRecursively.toSet).headOption
+//      println(lcaMaybe)
 //      x.foreach(println)
 
-      println(deriveCommonSuperType(nodeInfos))
+//      println(findSharedRoot(nodeInfos))
+      println(neighborInfoWith(Seq(nodeA1, baseNodeA)).deriveNeighborNodeType)
 
     }
-
-    def deriveCommonSuperType(nodeTypes: Set[AbstractNodeType]): Option[AbstractNodeType] = {
-      if (nodeTypes.size == 1) {
-        Some(nodeTypes.head)
-      } else if (nodeTypes.size > 1) {
-        /** Trying to find common supertype. This is nontrivial and we're probably missing a few cases.
-          * Trying to at least keep it deterministic...
-          * Idea: take one nodeType and check if it's type or any of it's supertypes are declared in *all* other nodeTypes
-          * */
-
-        val sorted = nodeTypes.toSeq.sortBy(_.className)
-
-        val (first, otherNodes) = (sorted.head, sorted.tail)
-        allTypes(first).find { candidate =>
-          otherNodes.forall { otherNode =>
-            allTypes(otherNode).contains(candidate)
-          }
-        }
-      } else {
-        None
-      }
-    }
-
-    def allTypes(node: AbstractNodeType): Seq[AbstractNodeType] =
-      node +: node.extendzRecursively
 
 
     def neighborInfoWith(nodes: Seq[AbstractNodeType]): NeighborInfoForEdge =
